Refactor: Unified button components with new design system (btn variants)
