+++ libs/ui/common/src/index.ts
@@ -5,7 +5,6 @@ export const DropdownDialog = { ..._DropdownDialog };
 export * from "./lib/author/author-chip";
 export * from "./lib/alert/alert";
 export * from "./lib/button/button";
-export * from "./lib/button/icon-only-button";
 export * from "./lib/button/dropdown-button";
 export * from "./lib/chip/image-chip";
 export * from "./lib/collapsible/collapsible-box";
