+++ package.json
@@ -39,7 +39,7 @@
 		"@monaco-editor/react": "^4.6.0",
 		"@next-auth/prisma-adapter": "^1.0.7",
 		"@nrwl/next": "19.5.6",
-		"@prisma/client": "^5.19.1",
+		"@prisma/client": "^5.22.0",
 		"@socialgouv/matomo-next": "^1.9.0",
 		"@tailwindcss/forms": "^0.5.7",
 		"@tanstack/react-query": "^4.36.1",
